server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
boto3==1.34.122
python-dotenv==1.0.1